refactor: extract android manager tests
